# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GTCP — Gas Trading & Commercial Platform API
  description: |
    **Gastrans Serbia · Backend REST API**

    Система управления газовыми транзитными потоками, договорами мощности, биллингом и аукционами.

    ### Нормативная база
    - **CAM NC EU 2017/459** — механизмы распределения мощности
    - **АЕРС 05-145** — официальные тарифы Gastrans
    - **Gastrans Network Code** — NC Art. 5 (Credit Support), Art. 6.3.1 (EXIT-SERBIA),
      Art. 8.3 (Surrender), Art. 18 (Fuel Gas), Art. 20 (Billing & Late Payment)
    - **MAR0277-24** — расписание аукционов CAM NC 2025-2026 (ENTSOG)

    ### Аутентификация
    Все endpoints (кроме `/auth/login`) требуют JWT Bearer token.

    ### Lifecycle аукциона
    ```
    Free Capacity → POST /auctions/bids (DRAFT)
                  → POST /auctions/bids/{id}/submit (SUBMITTED)
                  → POST /auctions/bids/{id}/result (WON/LOST)
                  → POST /auctions/bids/{id}/create-contract (CONTRACT_CREATED)
                  → POST /billing (Invoice)
    ```
  version: 5.0.0
  contact:
    name: GTCP Backend
    email: krav4ukleo@gmail.com
  license:
    name: Proprietary — Gastrans Serbia
    url: https://gastrans.rs
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development
  - url: https://gtcp.gastrans.rs/api/v1
    description: Production (Sprint 6)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token, срок действия 24 часа.
        Получить: `POST /auth/login` → `{ token }`.
...
security:
  - BearerAuth: []
paths:
...
  /nominations:
    get:
      tags: [nominations]
      summary: Список номинаций
      parameters:
        - in: query
          name: gas_date
          schema: { type: string, format: date }
        - in: query
          name: shipper_id
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, MATCHED, CONFIRMED, RENOMINATED, REJECTED] }
    post:
      tags: [nominations]
      summary: Создать номинацию
      description: ENTRY/EXIT матчинг ±10% tolerance, алгоритм min(entry, exit).
  /nominations/{id}/match:
    post:
      tags: [nominations]
      summary: Выполнить матчинг номинации
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
...